Move main node healtchcheck
