+++ core/bin/external_node/src/main.rs
@@ -35,7 +35,7 @@ use zksync_node_db_pruner::{DbPruner, DbPrunerConfig};
 use zksync_node_fee_model::l1_gas_price::MainNodeFeeParamsFetcher;
 use zksync_node_sync::{
     batch_status_updater::BatchStatusUpdater, external_io::ExternalIO,
-    tree_data_fetcher::TreeDataFetcher, ActionQueue, SyncState,
+    tree_data_fetcher::TreeDataFetcher, ActionQueue, MainNodeHealthCheck, SyncState,
 };
 use zksync_reorg_detector::ReorgDetector;
 use zksync_state::{PostgresStorageCaches, RocksdbStorageOptions};
@@ -54,7 +54,7 @@ use zksync_web3_decl::{
 
 use crate::{
     config::ExternalNodeConfig,
-    helpers::{MainNodeHealthCheck, ValidateChainIdsTask},
+    helpers::ValidateChainIdsTask,
     init::{ensure_storage_initialized, SnapshotRecoveryConfig},
     metrics::RUST_METRICS,
 };
