+++ core/node/node_sync/Cargo.toml
@@ -31,6 +31,7 @@ async-trait.workspace = true
 chrono.workspace = true
 tracing.workspace = true
 serde.workspace = true
+serde_json.workspace = true
 tokio = { workspace = true, features = ["time"] }
 thiserror.workspace = true
